Fix #12: Remove the banner no geofeature if event layer and geofeature are in events
