+++ src/app/modules/map/components/map-container/map-container.component.html
@@ -1,5 +1,5 @@
 <div class="map-container" [hidden]="!mapHasGeofeatures || !mapHasDataAnalytics">
-  <h5 class="text-center map-title">{{visualizationObject.mapConfiguration.name}}</h5>
+  <h6 class="text-center map-title">{{visualizationObject.mapConfiguration.name}}</h6>
   <div class="map-block__left">
     <ul class="map-controls__list">
       <li>
