+++ src/app/modules/map/containers/visualization-legend/visualization-legend.component.html
@@ -54,7 +54,7 @@
             </div>
           </div>
         </div>
-        <pagination-controls (pageChange)="handlePageChange($event)"></pagination-controls>
+        <pagination-controls (pageChange)="handlePageChange($event)" *ngIf="visualizationLegends.length > 3"></pagination-controls>
       </div>
       <div class="map-legend-list">
         <div class="map-legend-item">
