+++ src/app/modules/map/utils/index.ts
@@ -2,3 +2,4 @@ export * from './transformVisualizationObject';
 export * from './generateDhisUid';
 export * from './mapVisualization';
 export * from './transformFavourites';
+export * from './layers';
